requirements.txt
# Learnova LMS - Dependencies
Django>=5.0,<6.0
python-dotenv>=1.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
Pillow>=10.0
